Changed classes breakout, config and created class menu
