--- brick.py
@@ -1,13 +1,0 @@
-import pygame
-
-from game_object import GameObject
-
-
-class Brick(GameObject):
-    def __init__(self, x, y, w, h, color, special_effect=None):
-        GameObject.__init__(self, x, y, w, h)
-        self.color = color
-        self.special_effect = special_effect
-
-    def draw(self, surface):
-        pygame.draw.rect(surface, self.color, self.bounds)
