+++ config.py
@@ -48,7 +48,7 @@ button_normal_back_color = colors.INDIANRED1
 button_hover_back_color = colors.INDIANRED2
 button_pressed_back_color = colors.INDIANRED3
 
-menu_offset_x = 20
-menu_offset_y = 300
+menu_offset_x = 350
+menu_offset_y = 250
 menu_button_w = 80
 menu_button_h = 50
